requirements.txt
connexion
connexion[swagger-ui]
python-etcd
pandas
openpyxl